refactor: validator를 사용하는 주체 클래스 내부로 이동
- Bridge를 생성할 때 숫자 범위를 validation 하는것이 적절하다고 판단
- 추가로 테스트가 용이해진다는 장점을 가짐
