+++ docs/README.md
@@ -47,3 +47,11 @@
 - [x] BridgeGameStatus 테스트
   - [x] 상태를 보고 게임의 성공, 실패 여부 확인 테스트
   - [x] 게임의 상태를 올바르게 출력하는지 여부 테스트
+
+> 리팩토링이 필요한 부분 정리
+- [x] BridgeSizeValidator 클래스 위치 수정
+  - 해당 validator를 사용하는 객체 내부로 이동
+- [ ] GameCommandValidator 위치 수정
+  - 해당 validator를 사용하는 객체 내부로 이동
+- [ ] MoveDirectionValidator 위치 수정
+  - 해당 validator를 사용하는 객체 내부로 이동
